services/comment-service/pyproject.toml
[tool.poetry]
name = "comment-service"
version = "0.1.0"
description = "Comment microservice for ProStream platform"
authors = ["Abdumalik-ProDev <abdumalikbakhtiyorovwork@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.2"
uvicorn = { extras = ["standard"], version = "^0.30.6" }

# Database
sqlmodel = "^0.0.8"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"

# Messaging
redis = "^5.0.8"
kafka-python = "^2.0.2"

# gRPC
grpcio = "^1.66.1"
grpcio-tools = "^1.66.1"

# Auth / Security
pydantic = "^1.10.17"   # Compatible with sqlmodel
passlib = "^1.7.4"
python-jose = "^3.3.0"

# Utils
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"